Add unique persistent id test, organize attribute release and nameid tests, consolidate common attribute/nameid code
